+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-14 17:24:57
-// Commit: refactor(state): extract reusable atomic widgets from state
+// Updated at: 2026-05-15 20:16:06
+// Commit: feat(state): integrate offline caching for state state
 
 class stateComponentState {
-  final String commitRef = '1152';
-  final String lastMessage = 'refactor(state): extract reusable atomic widgets from state';
-  final DateTime timestamp = DateTime.parse('2026-05-14T17:24:57+05:00');
+  final String commitRef = '1193';
+  final String lastMessage = 'feat(state): integrate offline caching for state state';
+  final DateTime timestamp = DateTime.parse('2026-05-15T20:16:06+05:00');
 }